Extract some reusable-looking functions
