+++ modules/ghlint-api/src/main/kotlin/net/twisterrob/ghlint/model/Step.kt
@@ -71,9 +71,3 @@ public sealed interface Step {
 		public val ref: String
 	}
 }
-
-public val Step.stepsBefore: List<Step>
-	get() = parent.steps.subList(0, index.value)
-
-public val Step.stepsAfter: List<Step>
-	get() = parent.steps.subList(index.value + 1, parent.steps.size)
